# [FEM.elastic] Some optimisation to make FastTetrahedralCorotational even faster