@fredroy 2022-04-25
The scene (with this FEM) is OK, so good to go for me